Improve decode() function
Current implementation minimize the number of statements at the cost of a lot of branching. The new implementation provides systematic decoding for all but the last 4 characters and keeps the current implementation for the last 4 characters.

This new implementation with minimal branching provides slight performance improvement (between 3% and 18% on clang and gcc depending on the option; cf. this benchmark https://gist.github.com/Bouska/396323fe4ae81eec699e7ed716874ca6) at the cost of a few repeated C++ lines.